third_party/blink/web_tests/fast/forms/week-multiple-fields/week-multiple-fields-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for week input fields from renderer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Function arguments are (value, min, max).
Week
PASS stepUp("2000-W05", "2000-W05", "2000-W40") is "2000-W06"
PASS stepDown("2000-W05", "2000-W05", "2000-W40") is "2000-W40"
PASS stepUp("2000-W05", "2000-W02", "2000-W05") is "2000-W02"
PASS stepDown("2000-W05", "2000-W02", "2000-W05") is "2000-W04"
PASS stepUp("2000-W05", "2000-W08", "2000-W10") is "2000-W08"
PASS stepDown("2000-W05", "2000-W08", "2000-W10") is "2000-W10"
PASS stepUp("2000-W05", "2000-W03", "2000-W03") is "2000-W03"
PASS stepDown("2000-W05", "2000-W03", "2000-W03") is "2000-W03"
PASS stepUp("2000-W05", "2000-W03", "2001-W05") is "2000-W06"
PASS stepDown("2000-W05", "2000-W05", "2001-W08") is "2000-W04"
Year
PASS stepUp("2000-W01", "2000-W01", "2009-W52") is "2001-W01"
PASS stepDown("2000-W01", "2000-W01", "2009-W52") is "2009-W01"
PASS stepUp("2000-W01", "1990-W01", "2000-W52") is "1990-W01"
PASS stepDown("2000-W01", "1990-W01", "2000-W52") is "1999-W01"
PASS stepUp("275760-W01", "1990-W01", null) is "1990-W01"
PASS stepDown("1990-W01", "1990-W01", null) is "275760-W01"
PASS stepUp("2000-W01", null, "2000-W52") is "0001-W01"
PASS stepDown("0001-W01", null, "2000-W52") is "2000-W01"
PASS stepUp("1990-W01", "2000-W01", "2009-W52") is "2000-W01"
PASS stepDown("1990-W01", "2000-W01", "2009-W52") is "2009-W01"
PASS stepUp("2020-W01", "2000-W01", "2000-W52") is "2000-W01"
PASS stepDown("2020-W01", "2000-W01", "2000-W52") is "2000-W01"


PASS successfullyParsed is true

TEST COMPLETE
